Add a Makefile that allows to compile without Eclipse
This Makefile is really not ideal yet.

It has two target: compile, and run.

The compile target looks for all files in the src, test and web
directories, and recompiles everything if a single file changes.

This is currently done this way because we have no dependency
information from within make at the moment.

The run target will run a class (currently non configurable but we can
add some configuration to it). It depends on the compile target, so
everything will be recompiled each time a file is modified, even if it
is not a dependency of the target class.